+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Authentication management commands; see `reliably auth --help`
 - Help topic for environment variables descriptions; see `reliably environment --help`
+- Completion command for generating shell completion scripts; see `reliably completion --help`
 
 ## [0.2.0] - 2020-12-18
 ### Changed